[#17]: Allow to change exchange silver threshold
Indroduced options for actions.
It gives opportunity to specify custom options for each actions.
If there option `enable` and it is not `true`, then skip the action.

The code is not optimized for Actions with options.
In future actions should be extracted to separate classes,
and initialization of action should be done only once.

Update readme to mention about possobility to exchange the silver by
custom threshold. Changed servers examples to show this option.

The option enable is hidden for now.